Set up integration with Travis CI
Travis CI (travis-ci.org) is a service for automatically running tests
on pushes and pull requests. It is free for Open Source projects. It has
been enabled for the FeeLLGood GitHub repository.

This commit adds the infrastructure needed by Travis to do a test build:
 - .travis.yml is the Travis configuration file
 - install-dependencies.sh is a shell script for downloading, building
   and installing the required dependencies.
The commit also adds a small badge by the top of the README. The badge
should display "build passing", unless a recent commit broke the build.

The build logs can be found at https://travis-ci.com/feellgood/FeeLLGood