Save data with a fixed time step
In settings["outputs"], the integer parameter "save_energies" is
replaced by the floating point parameter "evol time step", which is the
time step for saving the evolution parameters to the .evol file.

The "take_photo" parameter is now in units of that time step.

The integration time steps are adjusted in order to exactly hit the
times required by the .evol file.