Move the magnetization parser to separate files
The parser is based on exprtk, a header-only, template-rich library that
takes very long to compile. In order to avoid recompiling it too often,
we move it out of the Settings class into a class of its own, defined in
mag_parser.{h,cpp}. The exprtk-based implementation is hidden from the
interface using the pimpl (Pointer to IMPLementation) idiom. Now
mag_parser.cpp is the only file which includes <exprtk.hpp>.