Optimize build time
The build time on Travis is optimized by letting `make' run multiple
jobs in parallel.

Since the compilation of mag_parser.cpp takes most of the build time, it
should be started first. It appears the sources are compiled in the
order they are listed in the `SOURCES' cmake variable. We then sort this
list in order of decreasing compilation time.