BUGFIX: inconsistent use of time step
In time_integration(), the time step dt was modified in multiple places
within the code. The value used to increment the physical time could
then be different from the value used to update the system state.

This commit introduces a new class, TimeStepper, that keeps all the
logic for updating the time step. Now, in time_integration(), dt is
updated only in a single place.