Fine tune stepper algorithm
Time is step in two nested loops: the inner loop goes through the times
that should be output to the file, while the inner loop uses adaptive
stepping.

The inner steps were previously constrained to evenly divide the
distance to the next outer loop time, and a "soft maximum" was always
increased irrespective of whether it was tested in an actual step.
Now the constraint to evenly divide the distance has been removed, and
the soft maximum is only increased if it has been actually tested.