Tune unit test numeric tolerances
Repeating `make test' a huge number of times revealed that,
occasionally, the errors are above the numerical tolerance. Analysis of
the tails of the distributions of rounding errors suggested updated
tolerances for the tests node_e_p, node_evol, and Tet_aniso_uniax.

For pt3D_triple_product, however, the errors can get quite large. This
test evaluates relative errors, but the absolute errors scale like a
quantity that can be arbitrarily larger than the numbers being compared.
We fix this by dividing the errors by the appropriate scaling factor,
and evaluating the absolute error in the quotient.

Fixes #8.