CI: Use explicit SHA1 for exprtk and ScalFMM
Both of these libraries suffer from weak version handling. If we want to
use versions that are both fixed (in order to get reproducible builds)
and not too old, we have to select specific SHAs from their respective
repos.

ScalFMM made no release since V1.5.1 (2018-04-11). We switch to using
the latest commit in the branch maintenance/scalfmm-1.5. This fixes a
couple of issues[1,2] for which we can now drop the patches we used to
carry in install-dependencies.sh.

Exprtk ostensibly ignores versioning: its version numbering is a joke,
the contents of the official zip download changes without notice (while
keeping the same file name), and the history on the GitHub repo gets
occasionally rewritten with fake commit dates.

Rather than selecting ourselves a specific exprtk commit, we follow the
choice of FreeBSD ports: they created their own versioning scheme for
exprtk, and they keep a well-versioned record of their version number
and the corresponding SHA1 in exprtk's GitHub repo.[3] We now use the
latest such version (g20210303). These commits usually get orphaned
soon, but GitHub does not seem to ever garbage-collect them. Just in
case it ever does, we now keep a fork with tags pointing to the commits
we use.[4]

[1] https://gitlab.inria.fr/solverstack/ScalFMM/-/issues/21
[2] https://gitlab.inria.fr/solverstack/ScalFMM/-/issues/22
[3] https://cgit.freebsd.org/ports/tree/math/exprtk/Makefile
[4] https://github.com/feellgood/exprtk