Fix lambda parameter in call of std::accumulate()
The first parameter of the lambda provided to std::accumulate() should
be passed either by value, or by const reference, but not as a simple
reference.

For some reason, gcc does not complain when the parameter is passed by
reference, if the code is being compiled as C++14 or C++17. When
compiling as C++20 or C++23, however, gcc 11.2.0 generates this error:

    cannot bind non-const lvalue reference of type ‘double&’ to an
    rvalue of type ‘std::remove_reference<double&>::type’ {aka ‘double’}