Save the magnetization state on SIGTERM
If the process receives the SIGTERM signal (which means "please
terminate ASAP"), terminate the current integration step, save the
magnetization state, then exit.

The magnetization is saved in a file named

    <basename>_at_exit.sol

The simulation time of this snapshot may not be a multiple of
evol_time_step, as the integration time step is set adaptively.