Make trailing '/' in outputs.directory optional
In the configuration file, the parameter outputs.directory need not end
anymore with a trailing '/'. If missing, the '/' will be added by
Settings::read(), unless the parameter is empty. Thus,
Settings::r_path_output_dir will always either be empty or end with '/'.
An empty value is interpreted as synonym for ".", the current working
directory.

settingsMaker.py doesn't need to add the trailing '/' anymore.