Define default settings in an embedded YAML file
The feeLLGood default settings are now stored in the file
default-settings.yml. This file is embedded in the executable as a blob,
and is read by the Settings() constructor.

In this YAML file, both mesh.volume_regions and mesh.surface_regions
contain an object name __default__. These objects are placeholders for
the defaults applicable to all the user-defined regions. These
placeholders are stored in the Settings object with -1 as their region
number.