Recentering and STT must be explicitly enabled
The simple presence of a "recentering" field in the configuration file
is not enough to enable recentering. Instead, the boolean field
recentering.enable must be explicitly set to true.

In the same vein, the spin transfer torque must be explicitly enabled by
setting spin_transfer_torque.enable to true.

This is intended to make it easier to store the default configuration in
an embedded YAML document that would be parsed before the user-supplied
configuration file.