Display statistics at program termination
The statistics are displayed in this format:

    Time step statistics:

        time steps       count       dt [*]          dumax [*]
        ──────────────────────────────────────────────────────────
        successful        1395   1.04e-09 ± 0.27   8.64e-04 ± 0.29
        failed             189   1.56e-09 ± 0.15

        [*] ranges given as (geometric mean) ± (relative stddev)

    Computing time:

        total: 82.1519 s (1.3692 min)
        per time step: 0.0588902 s

The “relative stddev” is actually the standard deviation of the
logarithm. For narrow distributions, this is equivalent to a “relative”
standard deviation (relative to the mean), e.g. “± 0.15” means ±15%.
Using logarithms makes the statistic better behaved with wide
distributions, and more robust to outliers.

The warnings “solver warning #1: you might need to adapt time step,
max(du) [...]” have been removed: they made feeLLGood too verbose in a
scary way. The only useful information they carried (namely the number
of such warnings) is now displayed as the count of failed time steps.

On normal program termination, both the time step statistics and the
computing times are printed. If feeLLGood is stopped with SIGTERM, only
the time step statistics are shown.