Make better use of Doxygen commands
Doxygen 1.9.4 produces multiple warnings on tiny.h about return types
being documented on functions that do not return anything. In order to
avoid this warning on a function that returns data through an output
parameter, the parameter should be documented with the command

    \param[out]

rather than with \return.

Fixing this required some rewriting and expanding of the documentation.
Some function parameters had also to be renamed in order to be
consistent with their documented names.