Generate version identifier with git describe
Use the output of

    git describe --tags --always --dirty

as the version identifier. This makes it obvious when the current
version does not match the latest release number, and also makes clear
when the working copy is "dirty", i.e. has uncommitted modifications.

As this renders the full SHA1 redundant, do not record it in config.h.

Since a "dirty" version string is quite long, the banner displayed at
program startup has to be dynamically sized to fit it.