CI: show machine architecture in cache entry name
GitHub introduced a user interface for managing action caches[1], which
makes it possible to monitor cache usage and remove redundant cache
entries. For this to be effective, it would be useful to easily identify
the machine architecture each cache entry belongs to.

Change the cache entry naming scheme in order to use the actual
architecture name instead of a hash.

[1] https://github.com/feellgood/FeeLLGood/actions/caches