Measure time in seconds
Contrary to what is stated in the documentation, time in feeLLGood is
not measured in seconds. Instead, the "time" variables represent some
sort of "reduced time" which is

    gamma0 * t

in units of (A/m)^(-1).

As this is highly unintuitive for the uninitiated user, switch all
user-visible times to be in seconds. Internally, all times are switched
to seconds except in the core of the solver. The conversions between
real and reduced time happen in the following places:

 - in Tet::integrales():
    - the real time prm_t.get_dt() is multiplied by gamma0 in order to
      get the reduced time s_dt
    - the term s_dt*Hv[npi] is divided by gamma0 because Hv[] is
      relative to real time

 - in timing::calc_alpha_eff(), the real time dt is multiplied by gamma0
   in order to get the reduced time reduced_dt

 - in mesh::updateNodes(), the reduced velocity X[] is multiplied by
   gamma0 in order to get the real velocity (vp, vq).

Adjust the script ci-tests/full_test.py to match the new time unit.