Check for received signals on every time step
When a step is retried, either because the solver did not converge or
because the magnetization changed too much, the current loop iteration
is aborted by a `continue' statement. This has the ill effect of
skipping the code that, by the end of the loop, checks whether a signal
has been received. In the pathological case where no step ever succeeds,
signals are never checked and the program cannot be interrupted with
SIGINT (Ctrl-C from keyboard) nor SIGTERM.

Fix this by checking for these signals at the beginning of the loop,
rather than at the end. Also, factor this check into its own function.