Fix the version identifier when no tags are found
The GitHub action actions/checkout@v3 we use in the CI pipeline fetches
a copy of the repository without the tags. In this situation, the `git
describe' command in CMakeLists.txt returns only the abbreviated SHA1 of
the current commit, instead of the usual format

    <closest_tag_name>-<extra_commit_count>-g<abbreviated_SHA1>

Then, the string(SUBSTRING ...) command used to remove the initial "v"
from the tag name removes instead the initial character of the SHA1.

Fix this by removing a match to the regex "^v" instead of whatever the
initial character may be.