Adjust the default settings for times in seconds
The default settings for

 - outputs.evol_time_step
 - time_integration.final_time
 - time_integration.min(dt)
 - time_integration.max(dt)

were suitable for the reduced time, but they are way too large now that
times are given in seconds. Fix this by scaling them by factors of the
order of 1e-5. The factors are not quite 1/gamma0, as they are chosen to
give nice round numbers.