CI: shorten the time integration in full_test.py
Instead of waiting until the system has almost relaxed, stop the
integration after 20 ps, which is about 3/4 of a precession on the
YZ plane. This makes the final state (and thus the whole test) more
sensitive to the dynamics.

Change also the time step, min(dt) and max(dt) to round numbers,
and get rid of the scaling factor gamma0.