Let the user choose the random seed
The command-line option `--seed <value>' lets the user choose the random
seed used by feeLLGood for setting the node-local basis. If the option
is not provided, a true random seed is generated by std::random_device.
Whether the seed is user-selected or not, it is displayed when feeLLGood
starts.

This option should enable the user to reproduce a feeLLGood run using
the exact same sequence of pseudo-random draws as in a previous run.

Unfortunately, it appears this option does not guarantee reproducible
runs. For a given random seed, the sequence of random draws is indeed
reproducible, and this was tested by printing them. Fixing the random
seed in full_test.py, however, does not prevent the random failures we
currently experience. Most of the times the test succeeds. It fails when
GMM prints the message "Level 1 Warning in
/usr/local/include/gmm/gmm_solver_bicgstab.h, line 112: Bicgstab failed
to converge".

It is unclear whether the observed non-reproducibility comes from a
library (maybe GMM or ScalFMM) or from an uninitialized variable in
feeLLGood.