Remove parallelism from LinAlgebra::insertCoeff()
This is a template method that processes either the volume elements or
the surface elements. It assembles the system of linear equations that
will be further solved by GMM, by gathering the contributions of all the
mesh elements. It uses the following algorithm:

    # Try to handle some elements in parallel:
    in parallel, for each (element):
        if (could acquire global mutex):
            element.add_contribution_to_linear_problem()
            element.treated = true
            release global mutex

    # Handle sequentially the remaining elements:
    sequentially, for each (element):
        if (not element.treated):
            element.add_contribution_to_linear_problem()
            element.treated = true

This approach has two issues:

1. The mutex protecting the system of linear equations, which is needed
   to prevent a data race, completely defeats the parallelism.

2. The coefficients of the linear system are computed by adding the
   contributions from the individual elements in an unpredictable order.

The second point makes the results of the simulations non-reproducible.
This is because, unlike the addition of real numbers, floating point
addition is not associative. Then, the coefficients of the linear system
depend on the unpredictable order in which the contributions are added.
Even though the dependence is usually weak (as it is a difference in the
rounding errors), it does significantly affect the behavior of the
solver.

This patch removes the attempt at handling elements in parallel.
Benchmarking on a simulation with a 10k-node mesh showed no measurable
difference in speed. In combination with the --seed option, this patch
makes the simulations perfectly reproducible.