Move setting `final_time' to the section `outputs'
In the settings file, the parameters `evol_time_step' and `final_time',
together, define the set of times that will be recorded in the evolution
file. They should be kept together.

The `outputs' section is appropriate, as it defines the data that is
requested by the user. The section `time_integration', which currently
holds `final_time', is meant for the tuning parameters of the adaptive
stepping algorithm.