convert2vtk: check validity of non-comment lines
The script tools/convert2vtk exits with an error message if the second
line of its input file does not have enough data fields. This comes from
the assumption that the first line, and only the first line, is a
comment. Since commit 4b94d85900723f43d95bb055856e5b9339eb8e95, this
assumption is not valid, and the script produces an error on all files
created by feeLLGood.

Instead of the second line, check all the lines with valid data.