Move cmake_minimum_required() to the top of CMakeLists.txt
The cmake-commands(7) manual advises:

    Call the cmake_minimum_required() command at the  beginning  of  the
    top-level CMakeLists.txt file even before calling the project() com‐
    mand.  It is important to establish version and policy settings  be‐
    fore invoking other commands whose behavior they may affect.

Let's follow the advice.

Note that cmake_minimum_required() implicitly invokes
cmake_policy(VERSION), which sets all the policies introduced in or
before the specified version to NEW. As CMP0048 was introduced in
CMake 3.0, we no longer need to set it explicitly.

Suppresses a warning issued by CMake 3.26.3 on the GitHub Actions
runners:

    CMake Warning (dev) at CMakeLists.txt:5 (PROJECT):
    cmake_minimum_required() should be called prior to this top-level
    project() call.  Please see the cmake-commands(7) manual for usage
    documentation of both commands.