Implement multiple verbosity levels
FeeLLGood's verbose mode is hardly usable because GMM is extremely
noisy. Fix this by making GMM silent unless the option "-v" is repeated
in the command line.

This is done by making Settings::verbose an integer representing the
"verbosity level": the number of occurrences of "-v" in the command
line. We ask GMM to be noisy only at levels 2 or higher.