Examples: get Python modules from feellgood package
Do not add "../tools" to the system path, as no modules are there
anymore. Instead, import them with the "feellgood." prefix: they will
be found in /usr/local/lib/python3.*/dist-packages/feellgood.

Note that the modules must be installed (with `sudo make install`) for
the example scripts to find them.