meshMaker.py: make the console output nicer
All the make() methods now use the same format for reporting about the
generated mesh:

    Generated <filename>: <cnt> nodes, <cnt> tetrahedra, <cnt> triangles

except that Cuboid does not print the triangle count.

The reports are build using f-strings, which is a feature introduced in
Python 3.6. This should be safe, as Python 3.5 is not supported since
September 2020, and the oldest distribution we support (Ubuntu Focal)
ships Python 3.8.