Unhide the expression parser implementations
Commit 0cb7f83a9a8d4e306178169857b9eaf575728aa9 introduced the pimpl
(Pointer to IMPLementation) idiom in order to hide the parser
implementations from the rest of the code. This was meant to avoid
exposing the <exprtk.hpp> header to more than one compilation unit, as
this header makes the compilation extremely slow.

Now that exprtk has been replaced by Duktape, compilation time is not an
issue. Remove this idiom in order to simplify the code.