Do not use const double variables in a constexpr
A `constexpr` should be initialized with a constant expression. A
`const` variable, however, is not usable in a constant expression unless
it is of integral or enumeration type.

Ensure that any `double` variable that appears in a `constexpr` is
itself `constexpr`. This fixes the following and similar errors reported
by GCC 13:

    facette.h:29:15: error: the value of ‘Facette::u’ is not usable in a
    constant expression
    facette.h:20:14: note: ‘Facette::u’ was not declared ‘constexpr’