Set the precision of the sol files to 16 digits
The floating point precision of output streams defaults to 6 significant
digits. This is insufficient to monitor the final stages of the total
energy of a system relaxing towards equilibrium.

An experiment showed that 15 digits are barely enough to see the energy
fluctuations induced by the numerical noise in the relaxed state. We
then increase the output precision to 16 significant digits, which
roughly matches the resolution of `double` variables.