BUGFIX: Output the magnetizations in the right order
Commit d4113e14d3 "Sort the nodes along the long axis of the sample"
changed the order of the nodes within feeLLGood relative to the order in
the mesh file. When outputting a magnetization configuration, feeLLGood
now uses this internal order, which is inconsistent with the original
mesh file, and makes the .sol files unusable.

Fix this by using the inverse permutation for outputting the node
magnetizations.