Let the user define Bext as a function
As an alternative to to an array of expressions, allow Bext to be
defined as a JavaScript function expression. For example, a rotating
field can be written as:

    Bext: |
        function(t) {
            var a = 2.5e-3,
                f = 0.955e9,
                phase = 2 * PI * f * t;
            return [a * cos(phase), a * sin(phase), 0];
        }

To make this possible, VectorParser::push_function() has been made
public. It was also renamed set_function(), which mirrors
set_expressions() and more clearly matches its intended use.