Let the user define initial_magnetization as a function
Similar to the previous commit, allow initial_magnetization to be
defined as a JavaScript function expression. For example, an ansatz of a
vortex can be written as:

    initial_magnetization: |
      function(x, y, z) {
        var r = sqrt(x**2 + y**2),
            theta = atan2(y, x),
            r0 = 10e-9,
            tanh = function(x) { return (exp(2*x)-1)/(exp(2*x)+1); },
            alpha = PI/2 * tanh(r/r0);
        return [-sin(theta)*sin(alpha), cos(theta)*sin(alpha), cos(alpha)];
      }

The initial magnetization can also be defined as the name of a .sol
file. Since both this and a function expression are strings, feeLLGood
uses the following heuristic for deciding what a string may represent in
this context:

* if it contains both the word "function" and an opening brace ("{"), it
  is assumed to be a function expression

* otherwise it is assumed to be a file name.