Add a destructor to the class VectorParser
Running a feeLLGood executable built with -fsanitize=address results in
multiple memory leaks being reported. The leaks are not troublesome, as
they only involve a couple of objects that have almost the lifetime of
the process. Clean up nevertheless in order to avoid noise from the
address sanitizer.

To comply with the rule of three, declare the copy constructor and the
copy assignment operator as `delete`.