Add useful math functions to the JS interpreter
These are a few math functions that, although standard in modern
JavaScript, are not provided by Duktape, nor are they part of ES5:

* shifted exp and log: expm1, log1p
* hyperbolic: cosh, sinh, tanh
* inverse hyperbolic: acosh, asinh, atanh

Care has been taken to avoid overflow of intermediate expressions, and
to keep reasonable accuracy (typically better that 1e-8 relative) in
edge cases.

If more speed and/or accuracy is needed, these function could be
reimplemented in C++ and fed to the interpreter with
duk_push_c_function() / duk_push_c_lightfunc().