CI: install MKL on Ubuntu and Debian
Installing it on Rocky Linux would be more complex, as there it is not
packages by the distribution: we may need to get the full oneAPI Intel
framework just for this.

On Debian, we need to enable the "non-free" distro component in order to
get libmkl-dev. On both Debian and Ubuntu, we need to ask for the
"noninteractive" Debian frontend, otherwise apt-get gets stuck when the
configuration of MKL asks whether libmkl_rt.so should be the default
alternative to BLAS/LAPACK.