Fix node order when reading magnetization from file
Since commit d4113e14d3 "Sort the nodes along the long axis of the
sample", the order of the nodes in the mesh::node array differs from the
order used in files. This is properly accounted for when writing
magnetization configuration (*.sol) files, but not when reading back
those files as `initial_magnetization`.

Fix mesh::readSol() to account for the changed node order.