Use Intel's oneMKL if available
On Debian and Ubuntu, a version of the Intel® Math Kernel Library (MKL)
from 2020 can be found in the standard repositories (resp. non-free and
multiverse). The user may instead get MKL directly from Intel,[1] either
because they want the latest version, or because it cannot be found in
the standard repos (e.g. on Rocky Linux). Let CMakeLists.txt use this
upstream version if available.

The current version of MKL is now called “oneAPI Math Kernel Library”
(oneMKL). Unlike the versions shipped by Debian and Ubuntu, which work
flawlessly with pkg-config, this upstream version:

* installs everything in /opt/intel, which requires setting a few
  environment variables before building, and setting at least
  LD_LIBRARY_PATH before running feellgood

* provides a script[2], which should be sourced by the shell in order to
  set these variables

* provides broken pkg-config .pc files listing non-existent include
  directories, which trigger a cmake failure

* provides a valid MKLConfig.cmake file, for use with find_package()

With this commit, CMakeLists.txt tries to load MKL using find_package().
If that fails, it tries find_package(PkgConfig). If that succeeds, it
tries to load MKL using the PkgConfig package. The expected result is:

* If oneMKL is installed and the environment has been properly set up,
  find_package() should succeed.

* If oneMKL is installed but the environment has not been set up, cmake
  will fail on the broken .pc files.

* If oneMKL is not installed and the distro-provided MKL is installed,
  find_package() should fail and PkgConfig should succeed.

* If no MKL is installed, feeLLGood will simply not use it.

Note that a failure of find_package(PkgConfig) is no longer fatal.

Tested on Ubuntu 22.04 and Rocky Linux 9.

[1] https://www.intel.com/content/www/us/en/developer/tools/oneapi/onemkl-download.html
[2] /opt/intel/oneapi/setvars.sh