Implement docker image build and push action
In install-dependencies.sh, install wget also on Debian-like OSes, as it
is not included the docker base image ubuntu:22.04.

As the container may run a on processor different from the one used in
the build, do not compile for the “native” instruction set. Use AVX2
instead, which is ubiquitous enough.

Build feeLLGood in /root/src, as this the directory used to build the
dependencies in install-dependencies.sh.

Push the generated container image to the GitHub Container Registry.