CI: remove Vagrant machine ubuntu_focal
Now that 24.04 is out, remove the old Ubuntu 20.04 LTS "Focal Fossa"
Vagrant machine: we only support the last two Ubuntu LTS releases (22.04
and 24.04).

Also, update the comment on Ubuntu 24.04 LTS "Noble Numbat", which is
not a "future" release anymore.

Note that Canonical has stopped releasing official Vagrant boxes.[1] The
box ubuntu/noble64, which I downloaded as a "development branch" prior
to this decision, is still functional, but not available for download
anymore. The box bento/ubuntu-24.04 seems like a good alternative.

[1] https://documentation.ubuntu.com/public-images/en/latest/public-images-explanation/vagrant/