Add new setting initial_time
The initial time of the simulation is:

 * zero if the initial magnetization is provided by a function or a set
   of analytic expressions (including the default: [0, 0, 1])

 * the time recorded in the .sol file if such a file is used to set the
   initial magnetization

The latter is a sensible choice when the user wants to continue a
simulation that was previously interrupted. It is common, however, to
perform one simulation with a big alpha, only to get an initial, relaxed
state for a subsequent simulation. In this case, one would want the time
to be reset to zero for the second simulation.

Add a new parameter `initial_time` to the YAML settings file, to let the
user explicitly set the initial simulation time. If unset, the current
default behavior applies.