modified input yaml template default-settings.yml to add boundary conditions of the electrostatic problem to solve for STT
The values of the boundary conditions (V and J) are children of the
surface regions. The default values (null) mean “not applicable”. For
consistency with ‘initial_time’, serialize these null as empty values.

If both V and J get assigned non-null values on the same surface, report
this as an error.