Settings: ignore null and empty parameters
The documentation suggests using the output of

    feellgood --print-defaults

as a template, and removing the settings that are to be left at their
default values. It is then quite easy for the user to end up with an
empty node such as:

    outputs:
      # nothing here: every setting within `outputs` has been removed

As the YAML syntax makes this a `null` value, feeLLGood rejects the file
with:

    CONFIGURATION ERROR: outputs should be a map.

Even feeLLGood is prone to creating settings files it cannot itself
accept. For instance, both of these commands:

    feellgood --verify /dev/null
    feellgood --print-defaults | feellgood --verify -

yield empty mesh.volume_regions and mesh.surface_regions.

Let feeLLGood accept null parameters, and handle them as if they were
undefined (i.e. a null value means “keep the default”). This should
hopefully make `feellgood --verify` idempotent.