Minimize rounding errors on time steps
The times that are saved to the .evol file are computed with a loop that
has the form:

    for (double t = t_initial; t < t_final + t_step / 2; t += t_step)
        ...

As each increment of the time (t += t_step) may carry a rounding error,
this form of loop tends to pile up many such errors.

Avoid this by looping over an integer “step number”, as in:

    for (int step_nb = 0; step_nb <= step_count; step_nb++)
        {
        double t = t_initial + step_nb * t_step;
        ...
        }

This form avoids the pile-up of rounding errors. An experiment with

    outputs:
      evol_time_step: 5e-13
      final_time: 5e-9

shows that, without this change, the final time is off by 177 ULPs
(units in the last place). With this change applied, the final time (the
double closest to 5e-9) is hit exactly.