Read defaults from a user's configuration file
FeeLLGood settings are now read from three sources. In this order:

 1. compiled-in defaults from default-settings.yml
 2. a user's personal configuration file, if it exists
 3. the file provided as a command-line argument

Each source can override the values defined in the previous sources.

The user's personal configuration file is:

 * $XDG_CONFIG_HOME/feellgood/default.yml if the environment variable
   XDG_CONFIG_HOME exists and is not empty
 * $HOME/.config/feellgood/default.yml otherwise

The purpose of this file is to let the user define in a single place
default settings that are not problem-specific, such as the number of
threads to be used by the solvers. In the future, this file could also
hold metadata (author, affiliation, licence, etc.) that feeLLGood would
copy in its output data files.