Add new setting outputs.metadata
This optional setting is meant to let the user provide arbitrary
metadata that feeLLGood will copy to every output file.

The value of outputs.metadata should be a YAML object (a collection of
"key: value" pairs), where the values are strings. Other scalar types
will be implicitly converted to strings.

For example, given a settings file with the following:

    outputs:
      metadata:
        Author: John Doe
        Affiliation: Center for the Study of Magic
        Sample: magnetic wand

feeLLGood will create output .evol and .sol files with the following
headers:

    ## Author: John Doe
    ## Affiliation: Center for the Study of Magic
    ## Sample: magnetic wand

This is in addition to the metadata that is already being saved in those
files ("feeLLGood version", "hostname", etc.).

If an item's value contains embedded newlines, these will be folded
RFC822-style, with a "##  " prefix (note the double space) on
continuation lines.

Note that metadata items that are not problem-specific, such as "Author"
and "Affiliation", can be conveniently stored in the user's
default-settings file ($HOME/.config/feellgood/default.yml).