Settings: use null instead of .nan
default-settings.yaml uses NaN values to represent “not applicable”,
e.g. the electrostatic potential on a surface with a boundary condition
that is not a fixed potential. However, NaN cannot be represented in
JSON syntax, which is an issue because we use JSON quite a bit in the
examples we provide.

In order to keep the data model JSON-compatible, use null instead of NaN
to mean “not applicable”. For consistency with ‘initial_time’, serialize
these null as empty values.

While we are at it, add some error checking: it is an error to assign
non-null values to both V and J on the same surface.