Use precomputed volumes for computing averages
The per-region and global averaged quantities are computed by
integrating both the quantity to be averaged and the volume of the
integration region. The latter is either a region volume or the total
mesh volume, and all of these are already known.

Use the known volumes instead of recomputing them each time.