fg-merge-sol: process general .sol files
FeeLLGood can output two sorts of .sol files:
 - magnetization snapshots
 - electrostatic potentials
So far, fg-merge-sol can only process the first kind.

Make fg-merge-sol more general and let it process any kind of .sol file.
Its only assumptions are that the first column of the file is the node
index, and that the lines are sorted by this index. All the columns
after the node index will be copied to the output file.

BREAKING CHANGE: when run on a magnetization snapshot, fg-merge-sol will
now output an extra column with the magnetic potential, as this is
present in the input file.