Build the list of all the mesh edges
These are the edges of all the tetrahedrons, i.e. the pairs of adjacent
nodes. This will be useful for computing the maximum angle between the
magnetizations of adjacent nodes.

The list is implemented as an std::vector with unique elements rather
than an std::set. Experimenting with a mesh having 10125 nodes,
51583 tetrahedrons and 65301 unique edges showed that the list is built
as an std::vector in 8 ms, whereas as an std::set it takes 44 ms.
Traversing the list is also expected to be faster with an std::vector.
The drawback is that, during construction, the std::vector consumes
48 bytes times the number of tetrahedrons.