Print the spin angle at program termination
Keep track of the maximum angle between the magnetizations of adjacent
nodes, where “maximum” is understood both across space and time. Print
this out at program termination, as a way to asses the quality of the
spatial sampling of the magnetization field.

Experimenting with the simulation of a vortex on a 100k-node mesh
showed the extra time required for this computation is too small
to be measurable.