install-dependencies: add python3-{gmsh,numpy}
full_test.py now depends on meshMaker, which itself depends on gmsh and
numpy. Install these python packages on Debian and Ubuntu.

Rocky Linux does have python3-numpy, but no python3-gmsh, so we install
neither of them. There is a python3-gmsh RPM packages in Fedora, but I
don't know whether it is feasible to install it on Rocky Linux.

Fixes: ModuleNotFoundError: No module named 'numpy'
https://github.com/feellgood/FeeLLGood/actions/runs/15191787264/job/42726204531