BUGFIX: unparallelize volume computation
The volumes of the sample regions are computed by a parallel loop over
the tetrahedrons, where each iteration adds a contribution to the volume
of the relevant region. Each region volume is thus a piece of data
accessed read/write by as many iteration as there are tetrahedrons in
that region, which is a data race.

Instead of parallelizing over the tetrahedrons, we could parallelize
over the regions. However, since this computation is only done once and
consume negligible time, we won't bother.

Fix the data race by unparallelizing the loop.

Fixes: 88bf0fee4b "Let each volume region store its volume".