w_sparseVect: replace the tree by an std::map
The write-mode sparse vector w_sparseVect is implemented as an ordered
binary tree holding v_coeff objects (basically, (index, value) pairs).
Replace the tree by an std::map mapping the indices to the values.
According to cppreference.com, the map is likely implemented as a
self-balanced ordered binary tree which, save for the self-balancing
property, is not very different from the current implementation.
However, as it is readily available and reasonably optimized, there is
no point in reinventing the wheel.

This saves 166 lines of code, without making the program any slower.
Both versions were benchmarked with a 10,125-node mesh, over five
repeated runs. The measured execution times in seconds are:

                min    median     max
    -----------------------------------
    tree      47.1084  47.2827  47.6456
    std::map  46.4762  47.2241  47.3953