full_test: build a sane mesh
full_test.py uses the mesh example/ellipsoid.msh, which is pathological,
as the central node is connected to the 162 other nodes. As we are going
to switch to a diagonal preconditioner which will not cope well with
such a mesh, we need a better mesh.

Let full_test.py build a sane ellipsoid mesh using the meshMaker module,
instead of relying example/ellipsoid.msh. This also makes the test more
similar to usual problems.

Update ci-tests/README.md and .gitignore to account for the generated
mesh.