Store r_sparseVect as two lists: indices and values
r_sparseVect is stored as an std::vector of (index, value) pairs, where
each such pair takes 12 bytes of data and 4 bytes of padding. Avoid the
padding by instead using separate vectors for the indices and the
values.

The performance gain for feeLLGood is hard to see, but it is clear on a
synthetic test repeatedly doing matrix-vector products (matrix with 200k
lines and 32 non-zeros per line). The loop in r_sparseVect::dot() can be
implemented with indices, iterators or pointers, but that makes no
measurable difference. Let's then use indices, as this gives the
simplest code.